feat(mcp): drain knowledge tool domain into a scoped pack (EP-8DC217EB BET-4)
Move search_knowledge, search_knowledge_base, create_knowledge_article, and
flag_stale_knowledge out of the mcp-tools.ts executeTool switch into a scoped
knowledge-pack. Definitions + handlers relocated verbatim; grants mirror
agent-grants.ts TOOL_TO_GRANTS. Inline-case ratchet baseline shrunk 212 -> 208.

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the knowledge tool domain into a scoped pack; ratchet baseline shrunk; behavior-preserving relocation.
