+++ scripts/mcp-tool-pack-baseline.json
@@ -35,7 +35,6 @@
   "create_ea_relationship",
   "create_employee",
   "create_epic",
-  "create_knowledge_article",
   "create_licensing_readiness_issue",
   "create_marketing_asset_task",
   "create_marketing_automation_candidate",
@@ -63,7 +62,6 @@
   "extract_brand_design_system",
   "fetch_public_website",
   "find_related_tests",
-  "flag_stale_knowledge",
   "generate_code",
   "generate_codebase_manifest",
   "generate_custom_archetype",
@@ -165,8 +163,6 @@
   "schedule_release_bundle",
   "search_code_graph",
   "search_design_intelligence",
-  "search_knowledge",
-  "search_knowledge_base",
   "search_portfolio_context",
   "search_project_files",
   "search_public_web",
